+++ src/sites/link/filemedia.net.js
@@ -1,0 +1,11 @@
+_.register({
+  rule: {
+    host: /^filemedia\.net$/,
+  },
+  async ready () {
+    const script = $.searchFromScripts(/window.location.href\s?=\s?\("(http.+?)"\)/);
+    if (script.length > 1) {
+      await $.openLink(script[1]);
+    }
+  },
+});