+++ src/sites/link/gamemod.pro.js
@@ -1,0 +1,10 @@
+_.register({
+  rule: {
+    host: /^gamemod\.pro$/,
+    path: /^\/download-file\//,
+  },
+  async ready () {
+    const gp = $('#wait-done > p > a');
+    await $.openLink(gp.href);
+  },
+});