+++ src/sites/link/healthykk.com.js
@@ -1,0 +1,10 @@
+_.register({
+  rule: {
+    host: /^healthykk\.com$/,
+    path: /^\/wordpress\//,
+  },
+  async ready () {
+    const h = $('#content center button');
+    h.click();
+  },
+});