+++ src/sites/link/namiyt.com.js
@@ -1,0 +1,12 @@
+_.register({
+  rule: {
+    host: [
+      /^www\.namiyt\.com$/,
+      /^realsht\.mobi$/,
+    ],
+  },
+  async ready () {
+    const n = $('#section1 form input#section1');
+    n.click();
+  },
+});