+++ src/sites/link/shtme.co.js
@@ -1,0 +1,10 @@
+_.register({
+  rule: {
+    host: /^shtme\.co$/,
+    path: /^\/\w+/,
+  },
+  async ready () {
+    const s = $('.content div a button');
+    s.click();
+  },
+});