+++ src/sites/link/thinfi.js
@@ -7,3 +7,13 @@ _.register({
     await $.openLink(a.href);
   },
 });
+
+_.register({
+  rule: {
+    host: /^hello\.tribuntekno\.com$/,
+  },
+  async ready () {
+    const a = $('div p u b a');
+    await $.openLink(a.href);
+  },
+});